refactor: 레거시 ParticipantVideo를 제거하고 My/Remote 테스트로 분리
